Move storage types to apiimpl package.
